Cell2 - another approach
After discovering performance problem in an application, tried an
alternative approach: Cell2 is rewritten from scratch. API is *not*
compatible with the original one. The main difference is that I use a
global stack to keep track of the execution context. This seems to be
a lot faster than modifying the self-object, in particular since it
does not require prototype chaining. It also makes defining cell
formulas easier (use any function, no need for clumpsy reference marking).
On the downside, the new approach will probably not work well when some kind
of concurrency acurs.